More experimentation.
Define a narrow "deadband" around the to-hold target.  Within that band
target speed is set by the speed PID.  Outside that band speed is set
such that a constant acc/deceleration of 5m/s^2 will get us to the edge
of the deadband, limited by a max speed of 5m/s.